ethiquest_backend/requirements.txt
# FastAPI and ASGI serverfastapi==0.109.1uvicorn[standard]==0.27.0.post1python-multipart==0.0.9pydantic==2.6.1pydantic-settings==2.1.0email-validator==2.1.0.post1# Databasesqlalchemy==2.0.25alembic==1.13.1asyncpg==0.29.0  # Async PostgreSQL driverpsycopg2-binary==2.9.9  # PostgreSQL adapterredis==5.0.1# Authentication and Securitypython-jose[cryptography]==3.3.0  # JWT tokenspasslib[bcrypt]==1.7.4  # Password hashingpython-dotenv==1.0.0cryptography==41.0.7# AI Integrationanthropic==0.8.1  # Claude API clientopenai==1.11.1    # OpenAI API client (backup)tiktoken==0.5.2   # Token counting for AI# Caching and Performancecachetools==5.3.2aioredis==2.0.1msgpack==1.0.7# Utilitiespython-dateutil==2.8.2pytz==2024.1aiofiles==23.2.1aiohttp==3.9.3aiodns==3.1.1tenacity==8.2.3    # Retry logichttpx==0.26.0      # Async HTTP client# Logging and Monitoringloguru==0.7.2prometheus-client==0.19.0opentelemetry-api==1.22.0opentelemetry-sdk==1.22.0opentelemetry-instrumentation-fastapi==0.43b0# Development Toolsblack==24.1.1        # Code formattingisort==5.13.2        # Import sortingflake8==7.0.0        # Lintingmypy==1.8.0          # Type checkingpytest==7.4.4        # Testingpytest-asyncio==0.23.5  # Async testingpytest-cov==4.1.0    # Test coveragefaker==22.6.0        # Test data generationrequests==2.31.0     # HTTP client for testing# Documentationmkdocs==1.5.3mkdocs-material==9.5.3mkdocstrings[python]==0.24.0# Emailaiosmtplib==3.0.1jinja2==3.1.3        # Email templates# Productiongunicorn==21.2.0supervisor==4.2.5sentry-sdk==1.40.0# Optional - Developmentipython==8.18.1      # Enhanced Python REPLjupyter==1.0.0       # Notebook interfacedebugpy==1.8.0       # Debugging# Version Managementpip-tools==7.3.0     # Dependency management[requires]python_version = "3.11"